# api/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: dsgamer777@gmail.com
    name: Omado
    url: https://twitter.com/shallow_omado
  description: アカウント関連API
  license:
    name: GPLv3
    url: https://www.gnu.org/licenses/gpl-3.0.ja.html
  title: UsagiBooru Accounts API
  version: "2.0"
servers:
- description: Prod
  url: https://api.gochiusa.team
- description: Staging
  url: https://api-next.gochiusa.team
- description: Dev
  url: http://127.0.0.1:8080
security:
- Authorization: []
...
paths:
  /accounts/{accountID}:
    delete:
      description: |-
        アカウントを削除します
        (本人または管理者のみがリクエストできます)
      operationId: deleteAccount
      parameters:
      - description: 取得したいユーザーのアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      - description: 削除確認用パスワード
        explode: true
        in: query
        name: password
        required: false
        schema:
          type: string
        style: form
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: No Content
        "403":
          content:
            application/json:
              examples: {}
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
      summary: Delete account info
      tags:
      - accounts
    get:
      description: ユーザーのアカウント情報を取得します
      operationId: getAccount
      parameters:
      - description: 取得したいユーザーのアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccountStruct'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
      summary: Get account info
      tags:
      - accounts
    patch:
      description: |-
        アカウント情報を編集します
        (本人または管理者のみがリクエストできます)
      operationId: editAccount
      parameters:
      - description: 取得したいユーザーのアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            examples:
              example:
                value:
                  userDisplayID: omado
                  userName: お窓
                  userPassword: kafuuch1n0
                  userOldPassword: h0t0c0c0a
                  userFavorite: 1
                  userTheme: 0
                  userPermission: 0
                  userLineID: dosugamea
                  userTwitterID: shallow_omado
            schema:
              $ref: '#/components/schemas/AccountStruct'
        description: |-
          ユーザー本人、もしくは管理者のみデータを変更できます
          変更したいカラムを与え、値に変更後の値を与えると変更できます
          パスワードを変更する場合のみ、例外的に現在のパスワードが要求されます
      responses:
        "200":
          content:
            application/json:
              examples: {}
              schema:
                $ref: '#/components/schemas/AccountStruct'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
      summary: Edit account info
      tags:
      - accounts
  /accounts/{accountID}/upload_history:
    get:
      description: イラスト投稿履歴を取得します
      operationId: getUploadHistory
      parameters:
      - description: 取得したいユーザーのアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      - description: 'ページ番号 (NOTE: 1ページ辺りの応答は20件)'
        explode: true
        in: query
        name: page
        required: false
        schema:
          minimum: 1
          type: integer
        style: form
      - description: ソート方法 d:最終更新
        explode: true
        in: query
        name: sort
        required: false
        schema:
          enum:
          - d
          example: d
          maxLength: 1
          minLength: 1
          type: string
        style: form
      - description: ソート方向 d:降順(大→小) a:昇順(小→大)
        explode: true
        in: query
        name: order
        required: false
        schema:
          default: d
          enum:
          - d
          - a
          example: d
          maxLength: 1
          minLength: 1
          type: string
        style: form
      - description: 1ページ辺りの要素数
        explode: true
        in: query
        name: per_page
        required: false
        schema:
          default: 20
          enum:
          - 20
          - 40
          - 60
          - 80
          - 100
          example: 20
          maximum: 100
          minimum: 20
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetUploadHistoryResponse'
          description: OK
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
      summary: Get upload history
      tags:
      - accounts
...
  /accounts/login/form:
    post:
      description: IDとパスワードを用いてユーザー認証し、APIトークンを発行します
      operationId: loginWithForm
      requestBody:
        content:
          application/json:
            examples:
              user:
                value:
                  id: test_user
                  password: this_is_only_for_test
              admin:
                value:
                  id: test_admin
                  password: this_is_only_for_test
            schema:
              $ref: '#/components/schemas/PostLoginWithFormRequest'
        description: ユーザーIDとパスワードをパラメータに取ります
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PostLoginWithFormResponse'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Bad Request
      security: []
      summary: Login with form
      tags:
      - accounts
...
  /accounts/{accountID}/mutes:
    get:
      description: 指定したアカウントのユーザーのミュート一覧を取得します
      operationId: getMutes
      parameters:
      - description: 対象のアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GetMutesResponse'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
      summary: Get mute list
      tags:
      - mutes
    post:
      description: 指定したアカウントのユーザーのミュートを追加します
      operationId: addMute
      parameters:
      - description: 対象のアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MuteStruct'
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MuteStruct'
          description: OK
        "400":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Bad Request
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Forbidden
        "429":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Too Many Requests
      summary: Add mute
      tags:
      - accounts
  /accounts/{accountID}/mutes/{muteID}:
    delete:
      description: 指定したミュート情報を削除します
      operationId: deleteMute
      parameters:
      - description: 対象のアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      - description: 対象のミュートID
        explode: false
        in: path
        name: muteID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "204":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: No Content
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
      summary: Delete mute
      tags:
      - accounts
    get:
      description: 指定したミュート情報を取得します
      operationId: getMute
      parameters:
      - description: 対象のアカウントID
        explode: false
        in: path
        name: accountID
        required: true
        schema:
          type: integer
        style: simple
      - description: 対象のミュートID
        explode: false
        in: path
        name: muteID
        required: true
        schema:
          type: integer
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MuteStruct'
          description: OK
        "403":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Forbidden
        "404":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GeneralMessageResponse'
          description: Not Found
      summary: Get mute
      tags:
      - mutes
components:
...
  schemas:
    AccountStruct:
      description: アカウントの取得/編集に使うリクエスト
      example:
        access:
          canCreatePost: true
          canApprovePost: false
          canInvite: true
          canLike: true
          canEditPost: false
          canComment: true
        apiKey: DUMMY_API_KEY
        apiSeq: 1
        ipfs:
          pinEnabled: false
          gatewayUrl: https://cloudflare-ipfs.com
          nodeEnabled: false
          nodeUrl: nodeUrl
          gatewayEnabled: false
        oldPassword: h0t0c0c0a
        description: 開発者さん
        permission: 0
        totpEnabled: false
        notify:
          hasWebNotify: false
          hasLineNotify: false
        accountID: 1
        password: kafuuch1n0
        name: お窓
        inviter:
          accountID: 1
          name: お窓
        invite:
          invitedCount: 0
          code: USAGIDEV
          inviteID: 1
        displayID: omado
        favorite: 1
      properties:
        accountID:
          description: ユーザーID
          example: 1
          type: integer
        displayID:
          description: (Twitterのような)表示IDを指定します。ここで指定したIDがログインに使用されます。英数字のみ入力できます。
          example: omado
          maxLength: 20
          minLength: 3
          type: string
        apiKey:
          description: APIキー
          example: DUMMY_API_KEY
          type: string
        apiSeq:
          description: 長期間有効トークン検証用シーケンス
          example: 1
          minimum: 1
          type: integer
        permission:
          default: 0
          description: 権限レベル 0:普通 5:Modelator 9:SysOp
          example: 0
          minimum: 0
          type: integer
        password:
          description: 新しいパスワードを入力します
          example: kafuuch1n0
          format: password
          maxLength: 50
          minLength: 8
          type: string
        oldPassword:
          description: |-
            現時点のパスワードを入力します。
            userPasswordを変更する場合に必要となります。
          example: h0t0c0c0a
          format: password
          maxLength: 50
          minLength: 8
          type: string
        totpEnabled:
          default: false
          description: TOTPが有効かが入ります
          type: boolean
        name:
          description: 他のユーザーに表示されるユーザー名/投稿者名
          example: お窓
          maxLength: 20
          minLength: 1
          type: string
        description:
          description: 他のユーザーに表示されるユーザー説明文/投稿者説明
          example: 開発者さん
          type: string
        favorite:
          description: ユーザーの推しキャラ(タグID)を選択します
          example: 1
          minimum: 0
          type: integer
        access:
          $ref: '#/components/schemas/AccountStruct_access'
        inviter:
          $ref: '#/components/schemas/LightAccountStruct'
        invite:
          $ref: '#/components/schemas/AccountStruct_invite'
        notify:
          $ref: '#/components/schemas/AccountStruct_notify'
        ipfs:
          $ref: '#/components/schemas/AccountStruct_ipfs'
      title: AccountStruct
      type: object
      x-examples:
        example-admin:
          accountID: 1
          displayID: omado
          apiKey: DUMMY_ADMIN_API_KEY
          apiSeq: 1
          permission: 9
          password: kafuuch1n0
          oldPassword: h0t0c0c0a
          totpEnabled: false
          name: お窓
          description: 開発者さん
          favorite: 1
          access:
            canInvite: true
            canLike: true
            canComment: true
            canCreatePost: true
            canEditPost: true
            canApprovePost: true
          inviter:
            accountID: 1
            name: お窓
          invite:
            inviteID: 1
            code: USAGIDEV
            invitedCount: 0
          notify:
            hasLineNotify: false
            hasWebNotify: false
          ipfs:
            gatewayUrl: https://cloudflare-ipfs.com
            nodeUrl: ""
            gatewayEnabled: false
            nodeEnabled: false
            pinEnabled: true
        example-user:
          accountID: 2
          displayID: chino
          apiKey: DUMMY_API_KEY
          apiSeq: 1
          permission: 0
          password: h0t0c0c0a
          oldPassword: kafuuch1n0
          totpEnabled: false
          name: 香風智乃
          description: ラビットハウスの一人娘。クールな性格だがココアのことは嫌いじゃない…!? コーヒーに詳しいしっかり者。
          favorite: 1
          access:
            canInvite: true
            canLike: true
            canComment: true
            canCreatePost: true
            canEditPost: false
            canApprovePost: false
          inviter:
            accountID: 1
            name: お窓
          invite:
            inviteID: 1
            code: USAGIDEV
            invitedCount: 0
          notify:
            hasLineNotify: false
            hasWebNotify: false
          ipfs:
            gatewayUrl: https://cloudflare-ipfs.com
            nodeUrl: ""
            gatewayEnabled: false
            nodeEnabled: false
            pinEnabled: false
    GeneralMessageResponse:
      example:
        message: message
      properties:
        message:
          description: 応答メッセージ(何が間違っているかの詳細)
          minLength: 1
          type: string
      required:
      - message
      title: GeneralMessageResponse
      type: object
      x-examples:
        not-found:
          message: Specified content was not found.
        no-permission:
          message: You don't have enough permission to do it.
        bad-request:
          message: Your request body is wrong.
        locked:
          message: Specified content is referenced from other content.
    GetUploadHistoryResponse:
      example:
        pagination:
          current: 1
          pages: 1
          perPage: 20
          count: 1
          title: 香風智乃
          type: tag
        contents:
        - artID: 1
          uploadID: 1
          started: 2021-02-21 22:24:22
          finished: 2021-02-21 22:24:27
          status: 5
        - artID: 1
          uploadID: 1
          started: 2021-02-21 22:24:22
          finished: 2021-02-21 22:24:27
          status: 5
      properties:
        pagination:
          $ref: '#/components/schemas/PaginationStruct'
        contents:
          description: |-
            指定された範囲で一致するデータ
            一致するものがなければ空配列
          items:
            $ref: '#/components/schemas/UploadHistoryStruct'
          minItems: 0
          type: array
          uniqueItems: true
      title: GetUploadHistoryResponse
      type: object
      x-examples:
        example:
          pagination:
            title: 投稿履歴一覧
            type: upload-history
            count: 1
            current: 1
            pages: 1
            perPage: 20
          contents:
          - uploadID: 1
            started: 2021-02-21 22:24:22
            finished: 2021-02-21 22:24:27
            status: 5
            artID: 1
    PostLoginWithFormRequest:
      example:
        password: password
        id: id
        totpCode: totpCode
      properties:
        id:
          description: ログインID
          minLength: 1
          type: string
        password:
          description: ログインパスワード
          minLength: 1
          type: string
        totpCode:
          description: ログインTOTPトークン
          type: string
      required:
      - id
      - password
      title: PostLoginWithFormRequest
      type: object
      x-examples:
        admin:
          id: omado
          password: kafuuch1n0
        user:
          id: chino
          password: h0t0c0c0a
    PostLoginWithFormResponse:
      example:
        apiKey: apiKey
      properties:
        apiKey:
          minLength: 1
          type: string
      title: PostLoginWithFormResponse
      type: object
      x-examples:
        example:
          apiKey: DUMMY_API_KEY
...
    GetMutesResponse:
      example:
        artists:
        - muteID: 1
          targetID: 1
          targetType: tag
        - muteID: 1
          targetID: 1
          targetType: tag
        tags:
        - muteID: 1
          targetID: 1
          targetType: tag
        - muteID: 1
          targetID: 1
          targetType: tag
      properties:
        tags:
          description: ミュート中のタグID配列
          items:
            $ref: '#/components/schemas/MuteStruct'
          type: array
        artists:
          description: ミュート中の絵師ID配列
          items:
            $ref: '#/components/schemas/MuteStruct'
          type: array
      required:
      - artists
      - tags
      title: GetMutesResponse
      type: object
      x-examples:
        example:
          tags:
          - muteID: 1
            targetType: tag
            targetID: 1
          artists: []
    MuteStruct:
      example:
        muteID: 1
        targetID: 1
        targetType: tag
      properties:
        muteID:
          description: ミュートID
          example: 1
          minimum: 1
          type: integer
        targetType:
          description: ミュート種別
          enum:
          - tag
          - artist
          example: tag
          type: string
        targetID:
          description: 対象のタグ/絵師ID
          example: 1
          minimum: 1
          type: integer
      title: MuteStruct
      type: object
      x-examples:
        example:
          muteID: 1
          targetID: 1
          targetType: tag
    LightAccountStruct:
      example:
        accountID: 1
        name: お窓
      properties:
        accountID:
          description: アカウントID
          example: 1
          minimum: 1
          type: integer
        name:
          description: アカウント名
          example: お窓
          type: string
      title: LightAccountStruct
      type: object
      x-examples:
        example:
          accountID: 1
          name: お窓
    PaginationStruct:
      example:
        current: 1
        pages: 1
        perPage: 20
        count: 1
        title: 香風智乃
        type: tag
      properties:
        title:
          description: ページネーションタイトル(表示用)
          example: 香風智乃
          minLength: 1
          type: string
        type:
          default: tag
          description: ページネーション種別(表示用)
          enum:
          - artist
          - tag
          - uploader
          - notify-client
          - notify-condition
          - mute
          - follow
          - article
          - mylist-list
          - art
          - upload-history
          example: tag
          type: string
        count:
          default: 0
          description: ヒット総数
          example: 1
          minimum: 0
          type: integer
        current:
          default: 1
          description: 現在のページ
          example: 1
          minimum: 1
          type: integer
        pages:
          default: 1
          description: ページ数
          example: 1
          minimum: 1
          type: integer
        perPage:
          default: 20
          description: ページ毎の取得数
          enum:
          - 20
          - 40
          - 60
          - 80
          - 100
          example: 20
          format: int32
          maximum: 100
          minimum: 20
          type: integer
      required:
      - count
      - current
      - pages
      - perPage
      - title
      - type
      title: PaginationStruct
      type: object
      x-examples:
        example:
          title: 香風智乃
          type: tag
          count: 1
          current: 1
          pages: 1
          perPage: 20
    UploadHistoryStruct:
      description: アップロード履歴
      example:
        artID: 1
        uploadID: 1
        started: 2021-02-21 22:24:22
        finished: 2021-02-21 22:24:27
        status: 5
      properties:
        uploadID:
          description: 通し投稿履歴番号(インデックス用)
          example: 1
          format: int32
          minimum: 1
          type: integer
        started:
          description: データ登録処理開始時刻
          example: 2021-02-21 22:24:22
          maxLength: 19
          minLength: 19
          type: string
        finished:
          description: データ登録完了時刻
          example: 2021-02-21 22:24:27
          maxLength: 19
          minLength: 19
          type: string
        status:
          default: 0
          description: 登録処理結果 5:成功 9:内部エラー
          enum:
          - 0
          - 5
          - 9
          example: 5
          maximum: 9
          minimum: 0
          type: integer
        artID:
          description: 投稿に成功した場合入るID
          example: 1
          format: int32
          minimum: 1
          type: integer
      required:
      - artID
      - finished
      - started
      - status
      - uploadID
      title: UploadHistory
      type: object
      x-examples:
        example:
          uploadID: 12045
          started: 2021-02-21 22:24:22
          finished: 2021-02-21 22:24:27
          status: 5
          artID: 9381
...
    AccountStruct_access:
      description: 管理者権限とは別の細かな権限
      example:
        canCreatePost: true
        canApprovePost: false
        canInvite: true
        canLike: true
        canEditPost: false
        canComment: true
      properties:
        canInvite:
          default: true
          description: 招待できるか
          type: boolean
        canLike:
          default: true
          description: いいねできるか
          type: boolean
        canComment:
          default: true
          description: コメントできるか
          type: boolean
        canCreatePost:
          default: true
          description: 投稿できるか
          type: boolean
        canEditPost:
          default: false
          description: 投稿を編集できるか
          type: boolean
        canApprovePost:
          default: false
          description: 投稿を承認できるか
          type: boolean
      type: object
    AccountStruct_invite:
      description: 招待情報
      example:
        invitedCount: 0
        code: USAGIDEV
        inviteID: 1
      properties:
        inviteID:
          description: 招待通し番号
          example: 1
          type: integer
        code:
          description: 招待コード(shortuuid)
          example: USAGIDEV
          maxLength: 12
          minLength: 6
          type: string
        invitedCount:
          default: 0
          description: 招待した人数の累計(誰を招待したかは表示されない)
          example: 0
          minimum: 0
          type: integer
      type: object
    AccountStruct_notify:
      description: 通知クライアントを設定済みか
      example:
        hasWebNotify: false
        hasLineNotify: false
      properties:
        hasLineNotify:
          default: false
          type: boolean
        hasWebNotify:
          default: false
          type: boolean
      type: object
    AccountStruct_ipfs:
      description: IPFS設定
      example:
        pinEnabled: false
        gatewayUrl: https://cloudflare-ipfs.com
        nodeEnabled: false
        nodeUrl: nodeUrl
        gatewayEnabled: false
      properties:
        gatewayUrl:
          default: https://cloudflare-ipfs.com
          description: 使用する任意のゲートウェイアドレス
          example: https://cloudflare-ipfs.com
          format: uri-reference
          type: string
        nodeUrl:
          description: 使用する任意のノードアドレス
          format: uri-reference
          type: string
        gatewayEnabled:
          default: false
          description: IPFSゲートウェイを使用するか否か
          type: boolean
        nodeEnabled:
          default: false
          description: IPFSノードを使用するか否か
          type: boolean
        pinEnabled:
          default: false
          description: マイリストを自動Pinningするか
          type: boolean
      type: object
...
  securitySchemes:
    Authorization:
      scheme: bearer
      type: http