# gen/model_general_message_response.go
/*
 * UsagiBooru Accounts API
 *
 * アカウント関連API
 *
 * API version: 2.0
 * Contact: dsgamer777@gmail.com
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package openapi

type GeneralMessageResponse struct {

	// 応答メッセージ(何が間違っているかの詳細)
	Message string `json:"message"`
}